+++ lib/whois/record/parser/whois.nic.net.ng.rb
@@ -7,7 +7,7 @@
 #++
 
 
-require 'whois/record/parser/base_cocca'
+require 'whois/record/parser/base_cocca2'
 
 
 module Whois
@@ -15,11 +15,7 @@ module Whois
     class Parser
 
       # Parser for the whois.nic.net.ng server.
-      #
-      # @see Whois::Record::Parser::Example
-      #   The Example parser for the list of all available methods.
-      #
-      class WhoisNicNetNg < BaseCocca
+      class WhoisNicNetNg < BaseCocca2
       end
 
     end