+++ CHANGELOG.md
@@ -35,6 +35,8 @@
 
 - CHANGED: Updated whois.na-nic.com.na parser to the new response format.
 
+- CHANGED: Updated whois.nic.net.ng parser to the new response format.
+
 - FIXED: whois.domainregistry.ie should support status `Active - LOCKED`.
 
 - FIXED: whois.nic.uk fails to parse registrars without URL (GH-188).